+++ lib/core/mixins/standalone_editor.dart
@@ -219,6 +219,10 @@ mixin StandaloneEditorState<
             image: imageBytes,
             isTransformed: isTransformed,
             layers: layers ?? [],
+            originalImageSize: null,
+            temporaryDecodedImageSize: null,
+            bodySize: null,
+            editorSize: null,
           ),
         );
       }